conagua-climat-station-stats-extractor-main/requirements.txt
requests
unidecode
jsonargparse[signatures,urls]
geopandas
